Refactor: Move chat history logic to ChatHistoryScreen
This commit removes the ConversationsContext and integrates its logic directly into the ChatHistoryScreen component. This simplifies the application structure by consolidating related functionality.
